Make RFactorial.evaluate use a long

+++ src/functions/unitary/combo/RFactorial.java
@@ -28,7 +28,7 @@ public class RFactorial extends Factorial {
 		if (!Settings.enforceIntegerOperations)
 			throw new IllegalStateException("RFactorial cannot be used if Settings.enforceIntegerOperations is not enabled.");
 		int argument = MiscTools.toInteger(operand.evaluate(variableValues));
-		int prod = 1;
+		long prod = 1;
 		for (int i = 1; i <= argument; i++)
 			prod *= i;
 		return prod;
